Update timetableActions.js
Pop-up window was showing up because of "history is undefined" error which im not really sure what it means. There were som informations about this error regardig to Reacts routing but i found no relation to axios o post. And im not sure what and if we should do something with this error or prevent it somehow. As of now i found no way how to fix this error, and when back arrow is pressed after creating "change subject request", im navigated back to window where i choose personal number, which is correct in my opinion. I changed the catch block, so now it shows only for response error from server or bad request to server. Other errors will be printed on console for now.